add initialization hook to kill node.
this is necessary for torque nodes/jobs, because nodes/jobs might be
queued for a long period of time. after they get out of the queue
they try to connect to the driver, which might not be running anymore.
without his hook, nodes/jobs on torque are running and reconnecting
until they are kicked from the job list (depending on the torque queue).